Added button features, changed layout for images
Added elements for images but I cant get them to show up when I launch
the file from my local machine. Not sure why I will figure this out later.
I fixed the container for the button and image iside the header container.
styled a hover feature for the button and the links.